feat: zksync deployment using custom hardhat-deploy
