+++ src/deploy/deploy_factories.ts
@@ -1,18 +1,19 @@
-import { DeployFunction } from "hardhat-deploy/types";
+import { DeployFunction } from "@elvis-krop/hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
+import { getDeployer } from "../zk-utils/getDeployer";
 
 const deploy: DeployFunction = async function (
   hre: HardhatRuntimeEnvironment,
 ) {
-  const { deployments, getNamedAccounts } = hre;
+  const { deployments, getNamedAccounts, network } = hre;
   const { deployer } = await getNamedAccounts();
   const { deploy } = deployments;
 
   await deploy("GnosisSafeProxyFactory", {
-    from: deployer,
+    from: network.zksync ? getDeployer(hre).zkWallet.privateKey : deployer,
     args: [],
     log: true,
-    deterministicDeployment: true,
+    deterministicDeployment: !network.zksync,
   });
 };
